+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Default max-cpu app limit to 400m
 - [HELM] Bump minio version to v0.5.5
 - [server] Change the pull policy to always pull
+- Infer if an app is a cronjob by `cron` prefix on process type
 
 ### Added
 - Support Nginx as sidecar